Add ticket functionality

+++ TravelExperiences/Model/ThingsToDo.swift
@@ -5,7 +5,7 @@ struct ThingsToDo {
     let price: Double
     let images : [String]
     let description: String
-    let rating: Int
+    let rating: Float
     let contactNumber: String
     
     init() {
@@ -13,11 +13,11 @@ struct ThingsToDo {
         self.price = 0.0
         self.images = ["shinjuku.png"]
         self.description = ""
-        self.rating = 0
+        self.rating = 0.0
         self.contactNumber = "(999) 999-9999"
     }
     
-    init (name: String, price: Double, images: [String], description: String, rating: Int, contactNumber: String) {
+    init (name: String, price: Double, images: [String], description: String, rating: Float, contactNumber: String) {
         self.name = name
         self.price = price
         self.images = images
